% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdelt.R
\name{fdelt}
\alias{fdelt}
\title{fdelt}
\usage{
fdelt(n, R, j)
}
\arguments{
\item{n}{sample size}

\item{R}{correlation matrix}

\item{j}{inner component of normal likelihood}
}
\value{
proportional likelihood
}
\description{
likelihood for correlation parameter with a uniform prior
}
\examples{
example(fdelt(n = 100, R = matrix(diag(c(1,1,1))), j = 10))
}
