% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatedata.R
\name{generatedata}
\alias{generatedata}
\title{simulate data for simulations}
\usage{
generatedata(n, mu, psi2, psi1, omega1, omega2, sig, R)
}
\arguments{
\item{n}{sample size}

\item{mu}{mean of outcomes}

\item{psi2}{coefficient for X for S(1)}

\item{psi1}{coefficient for X for S(0)}

\item{omega1}{coefficient for X for T(0)}

\item{omega2}{coefficient for X for T(1)}

\item{sig}{variance parameter}

\item{R}{correlation matrix}
}
\value{
dataset
}
\description{
simulate data
}
\examples{
example(generatedata(n = 100, mu = c(0, 2, 3, 4), 
psi2 = 0, psi1 = 0, omega1 = 1, omega2 = 1, sig = 1))
}
