% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.R
\name{nearest}
\alias{nearest}
\title{nearest}
\usage{
nearest(x, xval)
}
\arguments{
\item{x}{vector of values}

\item{xval}{value to find the nearest value in 'x' to}
}
\value{
index of 'x' with the value nearest to 'xval'
}
\description{
nearest function from GenKern package (no longer available)
}
\examples{
example(nearest(1:10, 3))
}
