% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{run_sim}
\alias{run_sim}
\title{run the simulation}
\usage{
run_sim(SIM, ST, X, n, condindfit)
}
\arguments{
\item{SIM}{number of iterations of mcmc to run}

\item{ST}{dataset}

\item{X}{covariate}

\item{n}{sample size}

\item{condindfit}{fit using conditional independence assumption}
}
\value{
simulation results
}
\description{
run mcmc simulation
}
\examples{
example(run_sim(SIM = 1000, ST = STdat, X = X, n = 100))
}
