% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim_obsdata.R
\name{run_sim_obsdata}
\alias{run_sim_obsdata}
\title{run the simulation}
\usage{
run_sim_obsdata(SIM, ST, X, trt, condindfit, grid, min, max)
}
\arguments{
\item{SIM}{number of iterations of mcmc to run}

\item{ST}{dataset}

\item{X}{covariate}

\item{trt}{treatment vector}

\item{condindfit}{boolean for conditional independence}

\item{grid}{interval of values for assessing valid covariance matrices}

\item{min}{minimum value of correlations}

\item{max}{maximum value of correlations}
}
\value{
simulation results
}
\description{
run mcmc simulation
}
\examples{
example(run_sim(SIM = 1000, ST = STdat, X = X, n = 100))
}
