% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim_unequalarms.R
\name{run_sim_unequalarms}
\alias{run_sim_unequalarms}
\title{run the simulation}
\usage{
run_sim_unequalarms(SIM, ST, X, trt)
}
\arguments{
\item{SIM}{number of iterations of mcmc to run}

\item{ST}{dataset}

\item{X}{covariate}

\item{trt}{treatment allocation}
}
\value{
simulation results
}
\description{
run mcmc simulation
}
\examples{
example(run_sim(SIM = 1000, ST = STdat, X = X, n = 100))
}
