% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_prior.R
\name{uniform_prior}
\alias{uniform_prior}
\title{uniform_prior}
\usage{
uniform_prior(n, R, j)
}
\arguments{
\item{n}{sample size}

\item{R}{correlation matrix}

\item{j}{inner component of normal likelihood}
}
\value{
proportional likelihood
}
\description{
likelihood for correlation parameter with a uniform prior
}
\examples{
example(uniform_prior(n = 100, R = matrix(diag(c(1,1,1))), j = 10))
}
