% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prior.R
\name{beta_prior}
\alias{beta_prior}
\title{beta_prior}
\usage{
beta_prior(n, R, j, coord1, coord2)
}
\arguments{
\item{n}{sample size}

\item{R}{correlation matrix}

\item{j}{inner component of normal likelihood}

\item{coord1}{row coordinate of correlation matrix}

\item{coord2}{column coordinate of correlation matrix}
}
\value{
proportional likelihood
}
\description{
likelihood for correlation parameter with a Beta prior
}
\examples{
example(beta_prior(n = 100, R = matrix(diag(c(1, 1, 1))), j = 10))
}
