% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R, R/run_sim_covariates.R
\name{run_sim}
\alias{run_sim}
\title{run the simulation}
\usage{
run_sim(SIM, ST, X, n, condindfit, trt)

run_sim(SIM, ST, X, n, condindfit, trt)
}
\arguments{
\item{SIM}{number of iterations of mcmc to run}

\item{ST}{dataset}

\item{X}{covariate}

\item{n}{sample size}

\item{condindfit}{fit using conditional independence assumption}

\item{trt}{treatment allocation vector}
}
\value{
simulation results

simulation results
}
\description{
run mcmc simulation

run mcmc simulation
}
\details{
importFrom("stats", "coef", "complete.cases", "cor", "lm", "median",
              "na.fail", "na.omit", "quantile", "runif", "sd", "var", "utils", "write.table")

importFrom("stats", "coef", "complete.cases", "cor", "lm", "median",
              "na.fail", "na.omit", "quantile", "runif", "sd", "var", "utils", "write.table")
}
\examples{
example(run_sim(SIM = 1000, ST = STdat, X = X, n = 100))
example(run_sim(SIM = 1000, ST = STdat, X = X, n = 100))
}
